+++ configs/_base_/models/bsn_tem.py
@@ -6,6 +6,3 @@ model = dict(
     tem_feat_dim=400,
     tem_hidden_dim=512,
     tem_match_threshold=0.5)
-# model training and testing settings
-train_cfg = None
-test_cfg = dict(average_clips='score')