+++ docs/changelog.md
@@ -6,15 +6,21 @@
 
 **New Features**
 
-- Support LFB [#553](https://github.com/open-mmlab/mmaction2/pull/553)
+- Support LFB ([#553](https://github.com/open-mmlab/mmaction2/pull/553))
+- Support using backbones from MMCls for TSN ([#679](https://github.com/open-mmlab/mmaction2/pull/679))
 
 **Improvements**
+
+- Add slowfast config/json/log/ckpt for training custom classes of AVA ([#678](https://github.com/open-mmlab/mmaction2/pull/678))
 
 **Bug and Typo Fixes**
 
 **ModelZoo**
 
-- Add LFB for AVA2.1 [#553](https://github.com/open-mmlab/mmaction2/pull/553)
+- Add LFB for AVA2.1 ([#553](https://github.com/open-mmlab/mmaction2/pull/553))
+- Add slowonly_nl_embedded_gaussian_r50_4x16x1_150e_kinetics400_rgb ([#690](https://github.com/open-mmlab/mmaction2/pull/690))
+- Add TSN with ResNeXt-101-32x4d backbone ([#679](https://github.com/open-mmlab/mmaction2/pull/679))
+- Add slowonly_nl_embedded_gaussian_r50_8x8x1_150e_kinetics400_rgb ([#704](https://github.com/open-mmlab/mmaction2/pull/704))
 
 ### 0.12.0 (28/02/2021)